@@ -7,10 +7,10 @@
   "type": "module",
   "dependencies": {
     "@tmtsoftware/esw-ts": "",
-    "antd": "^4.16.5",
+    "antd": "^4.18.5",
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
-    "react-router-dom": "^5.2.0"
+    "react-router-dom": "^6.2.1"
   },
   "scripts": {
     "start": "vite",
@@ -26,31 +26,31 @@
     "last 1 chrome version"
   ],
   "devDependencies": {
-    "@testing-library/react": "^12.0.0",
-    "@testing-library/react-hooks": "^7.0.0",
-    "@testing-library/user-event": "^13.1.9",
-    "@types/chai": "^4.2.19",
-    "@types/react": "^17.0.11",
-    "@types/react-dom": "^17.0.8",
-    "@types/react-router-dom": "^5.1.7",
-    "@typescript-eslint/eslint-plugin": "^4.28.0",
-    "@typescript-eslint/parser": "^4.28.0",
-    "@web/test-runner": "^0.13.12",
-    "chai": "^4.3.4",
-    "eslint": "^7.29.0",
+    "@testing-library/react": "^12.1.2",
+    "@testing-library/react-hooks": "^7.0.2",
+    "@testing-library/user-event": "^13.5.0",
+    "@types/chai": "^4.3.0",
+    "@types/react": "^17.0.38",
+    "@types/react-dom": "^17.0.11",
+    "@types/react-router-dom": "^5.3.3",
+    "@typescript-eslint/eslint-plugin": "^5.10.1",
+    "@typescript-eslint/parser": "^5.10.1",
+    "@web/test-runner": "^0.13.26",
+    "chai": "^4.3.6",
+    "eslint": "^8.8.0",
     "eslint-config-prettier": "^8.3.0",
-    "eslint-import-resolver-typescript": "^2.4.0",
-    "eslint-plugin-import": "~2.23.4",
-    "eslint-plugin-prettier": "^3.4.0",
-    "eslint-plugin-react": "^7.24.0",
-    "eslint-plugin-react-hooks": "^4.2.0",
-    "eslint-plugin-testing-library": "^4.6.0",
+    "eslint-import-resolver-typescript": "^2.5.0",
+    "eslint-plugin-import": "~2.25.4",
+    "eslint-plugin-prettier": "^4.0.0",
+    "eslint-plugin-react": "^7.28.0",
+    "eslint-plugin-react-hooks": "^4.3.0",
+    "eslint-plugin-testing-library": "^5.0.4",
     "npm-run-all": "^4.1.5",
-    "prettier": "^2.3.1",
-    "vite": "^2.7.10",
+    "prettier": "^2.5.1",
+    "vite": "^2.7.13",
     "vite-web-test-runner-plugin": "^0.1.0",
     "@vitejs/plugin-react-refresh": "^1.3.6",
     "ts-mockito": "^2.6.1",
-    "typescript": "^4.3.4"
+    "typescript": "^4.5.5"
   }
 }
