@@ -1,5 +1,5 @@
 import React from 'react'
-import { Route, Switch } from 'react-router-dom'
+import { Route, Routes as RouterRoutes } from 'react-router-dom'
 import { NotFound } from '../components/error/NotFound'
 import { RaDec } from '../components/pages/RaDec'
 import { SecuredRaDecInput } from '../components/pages/SecuredRaDecInput'
@@ -7,14 +7,14 @@ import { ProtectedRoute } from './ProtectedRoute'
 
 export const Routes = (): JSX.Element => {
   return (
-    <Switch>
+    <RouterRoutes>
       {/*// #add-route */}
-      <Route exact path='/' component={RaDec} />
+      <Route path='/' element={<RaDec />} />
       {/*// #add-route */}
       {/*// #add-protected-route */}
-      <ProtectedRoute path='/securedRaDec' component={SecuredRaDecInput} />
+      <ProtectedRoute path='/securedRaDec' element={<SecuredRaDecInput />} />
       {/*// #add-protected-route */}
-      <Route path='*' component={NotFound} />
-    </Switch>
+      <Route path='*' element={<NotFound />} />
+    </RouterRoutes>
   )
 }
